microservicio-ia/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
google-generativeai==0.6.0
pydantic==2.9.2
